b486c83625e49eece0548c3db4ee3f84196987ea Use usize for CompactSizeEncoder::new (Mitchell Bagot)

Pull request description:

  The current CompactSizeEncoder::new() constructor uses an impl ToU64 for its first parameter. The ToU64 trait is part of internals, and thus isn't permitted to be part of the public API of consensus_encoding.
  
  Change the ToU64 parameter type to usize, encoding values outside of the u64 range (such as on 128 bit system) as u64::MAX.
  
  Closes #5343


ACKs for top commit:
  apoelstra:
    ACK b486c83625e49eece0548c3db4ee3f84196987ea; successfully ran local tests
  tcharding:
    ACK b486c83625e49eece0548c3db4ee3f84196987ea


Tree-SHA512: 0a7064e3485949414590002e6fbfb6325bb6a76ce759bd6f40f4c7c1ca4817613e51e3d344f8c2b967855d2ba9d430667141f8196561c107275cbe5a63ced5ca

+++ primitives/src/transaction.rs
@@ -330,11 +330,11 @@ impl Encodable for Transaction {
     fn encoder(&self) -> Self::Encoder<'_> {
         let version = self.version.encoder();
         let inputs = Encoder2::new(
-            CompactSizeEncoder::new(self.inputs.len() as u64),
+            CompactSizeEncoder::new(self.inputs.len()),
             SliceEncoder::without_length_prefix(self.inputs.as_ref()),
         );
         let outputs = Encoder2::new(
-            CompactSizeEncoder::new(self.outputs.len() as u64),
+            CompactSizeEncoder::new(self.outputs.len()),
             SliceEncoder::without_length_prefix(self.outputs.as_ref()),
         );
         let lock_time = self.lock_time.encoder();
